Setup the package for multiple exports
We want to leverage Javy's new multi-export feature. This commit helps
by:
1. Extracting the logic for running a function (stream & JSON handling)
   to a `run` function that can be used without `index.ts`.
2. Making `javy` a peerDependency so that it's still accessible for
   `run`, (without using any file here as esbuild's entrypoint).
